requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
httpx>=0.27.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
upstash-redis>=1.1.0
APScheduler>=3.10.4
